(trunk libT) fix minor compiler warning that cropped up when upgrading our ConvertUTF8

+++ libtransmission/bencode.c
@@ -1235,7 +1235,7 @@ jsonStringFunc( const tr_benc * val,
                     if( ( result != conversionOK ) && ( tmp == it ) )
                         ++it; /* it's beyond help; skip it */
                     else {
-                        evbuffer_add_printf( data->out, "\\u%04x", buf );
+                        evbuffer_add_printf( data->out, "\\u%04x", (unsigned int)buf );
                         it = tmp - 1;
                     }
                     /*fprintf( stderr, "[\\u%04x]\n", buf );*/
